+++ client/Projects/OpenList/OpenList/Scenes/TabbarScene/AddTabScene/AddCheckListTitleScene/AddCheckListTitleViewFactory.swift
@@ -67,15 +67,3 @@ final class AddCheckListTitleViewFactory: Factory<AddCheckListTitleDependency>, 
 		return viewController
 	}
 }
-
-final class AccessTokenInterceptor: RequestInterceptor {
-	public func intercept(_ request: URLRequest) -> URLRequest {
-		var request = request
-		if let accessToken = KeyChain.shared.read(key: AuthKey.accessToken) {
-			request.addValue("Bearer \(accessToken)", forHTTPHeaderField: "Authorization")
-			return request
-		} else {
-			return request
-		}
-	}
-}
